refactor: Flatten file structure and fix imports
This commit refactors the codebase to integrate adaptive superpixel
functionality directly into the existing `suit.py` and `utils.py` files,
removing the `models/` and `utils/` subdirectories that were
previously introduced. This addresses your feedback and resolves
`ModuleNotFoundError` issues related to package imports.

Key changes:

1.  **File Structure Flattening**:
    *   The `ParameterPredictor` class definition has been moved directly
        into `suit.py`.
    *   The `generate_superpixels` function and its `SKIMAGE_AVAILABLE`
        helper constant have been moved directly into `utils.py`.
    *   The `SuitAdaptive` model definition (previously
        `models/suit_adaptive.py`) now constitutes the main content of
        `suit.py`, replacing the original `SuperpixelVisionTransformer` logic.
        Registered model names (e.g., `suit_tiny_224_adaptive`) are
        preserved.
    *   The `models/` and `utils/` directories and their specific files
        (`parameter_predictor.py`, `suit_adaptive.py`, `superpixels.py`)
        have been removed.

2.  **Import Adjustments**:
    *   All import statements in `main.py`, `tests/test_adaptive.py`, and
        `visualize/adaptive_debug.py` have been updated to reflect the
        new locations of `ParameterPredictor` (in `suit.py`),
        `generate_superpixels` (in `utils.py`), and the `SuitAdaptive`
        model variants (in `suit.py`).
    *   Mock paths in tests have also been updated.

This resolves the previously reported `ModuleNotFoundError` and aligns
the codebase with the preferred flat directory structure.
